fix : refactor discord bot
